+++ src/main/java/ru/vyarus/dropwizard/guice/module/lifecycle/internal/LifecycleSupport.java
@@ -77,6 +77,8 @@ public final class LifecycleSupport {
     public void runPhase(final Configuration configuration,
                          final ConfigurationTree configurationTree,
                          final Environment environment) {
+        broadcast(new BeforeRunEvent(GuiceyLifecycle.BeforeRun, options, bootstrap,
+                configuration, configurationTree, environment));
         this.configuration = configuration;
         this.configurationTree = configurationTree;
         this.environment = environment;